Refactor: Move `social_django` factories to own dir.
That will distinguish them clearly from the ones that we define, avoiding some
possible confusion, clutter, and clashes.
