+++ tests/factories/__init__.py
@@ -6,7 +6,6 @@ from .job import *  # noqa: F401, F403
 from .job_request import *  # noqa: F401, F403
 from .org import *  # noqa: F401, F403
 from .org_membership import *  # noqa: F401, F403
-from .partial import *  # noqa: F401, F403
 from .project import *  # noqa: F401, F403
 from .project_collaboration import *  # noqa: F401, F403
 from .publish_request import *  # noqa: F401, F403
@@ -18,7 +17,8 @@ from .release_file_review import *  # noqa: F401, F403
 from .repo import *  # noqa: F401, F403
 from .site_alert import *  # noqa: F401, F403
 from .snapshot import *  # noqa: F401, F403
+from .social_django.partial import *  # noqa: F401, F403
+from .social_django.user_social_auth import *  # noqa: F401, F403
 from .stats import *  # noqa: F401, F403
 from .user import *  # noqa: F401, F403
-from .user_social_auth import *  # noqa: F401, F403
 from .workspace import *  # noqa: F401, F403
