+++ tests/factories/social_django/__init__.py
@@ -1,0 +1,2 @@
+"""Contains factories from social_django.models so that they are separated from
+the factories for models we define."""
